reorg code for builtins and intrinsics
