+++ src/Makefile
@@ -44,7 +44,7 @@ SRCS := \
 	dlload sys init task array dump toplevel jl_uv datatype \
 	simplevector APInt-C runtime_intrinsics runtime_ccall \
 	threadgroup threading stackwalk gc gc-debug gc-pages method \
-	jlapi signal-handling safepoint jloptions timing subtype
+	jlapi signal-handling safepoint jloptions timing subtype rtutils
 
 ifeq ($(USEMSVC), 1)
 SRCS += getopt
