+++ src/intrinsics.h
@@ -114,112 +114,3 @@ enum intrinsic {
 #undef ALIAS
     num_intrinsics
 };
-
-#ifdef __cplusplus
-extern "C"
-#endif
-JL_DLLEXPORT const char *jl_intrinsic_name(int f)
-{
-    switch ((enum intrinsic)f) {
-    default: return "invalid";
-#define ADD_I(func, nargs) case func: return #func;
-#define ADD_HIDDEN ADD_I
-#define ALIAS ADD_I
-    INTRINSICS
-#undef ADD_I
-#undef ADD_HIDDEN
-#undef ALIAS
-    }
-}
-
-static void (*runtime_fp[num_intrinsics])(void);
-static unsigned intrinsic_nargs[num_intrinsics];
-
-#define jl_is_intrinsic(v)       jl_typeis(v,jl_intrinsic_type)
-
-#ifdef __cplusplus
-extern "C"
-#endif
-JL_CALLABLE(jl_f_intrinsic_call)
-{
-    JL_NARGSV(intrinsic_call, 1);
-    JL_TYPECHK(intrinsic_call, intrinsic, F);
-    enum intrinsic f = (enum intrinsic)*(uint32_t*)jl_data_ptr(F);
-    if (f == cglobal && nargs == 1)
-        f = cglobal_auto;
-    unsigned fargs = intrinsic_nargs[f];
-    if (!fargs)
-        jl_error("this intrinsic must be compiled to be called");
-    JL_NARGS(intrinsic_call, fargs, fargs);
-
-    union {
-        void (*fptr)(void);
-        jl_value_t *(*call1)(jl_value_t*);
-        jl_value_t *(*call2)(jl_value_t*, jl_value_t*);
-        jl_value_t *(*call3)(jl_value_t*, jl_value_t*, jl_value_t*);
-        jl_value_t *(*call4)(jl_value_t*, jl_value_t*, jl_value_t*, jl_value_t*);
-    } fptr;
-    fptr.fptr = runtime_fp[f];
-    switch (fargs) {
-        case 1:
-            return fptr.call1(args[0]);
-        case 2:
-            return fptr.call2(args[0], args[1]);
-        case 3:
-            return fptr.call3(args[0], args[1], args[2]);
-        case 4:
-            return fptr.call4(args[0], args[1], args[2], args[3]);
-        default:
-            assert(0 && "unexpected number of arguments to an intrinsic function");
-    }
-    gc_debug_critical_error();
-    abort();
-}
-
-static void add_intrinsic_properties(enum intrinsic f, unsigned nargs, void (*pfunc)(void))
-{
-    intrinsic_nargs[f] = nargs;
-    runtime_fp[f] = pfunc;
-}
-
-static void add_intrinsic(jl_module_t *inm, const char *name, enum intrinsic f)
-{
-    jl_value_t *i = jl_box32(jl_intrinsic_type, (int32_t)f);
-    jl_sym_t *sym = jl_symbol(name);
-    jl_set_const(inm, sym, i);
-    jl_module_export(inm, sym);
-}
-
-#ifdef __cplusplus
-extern "C"
-#endif
-void jl_init_intrinsic_properties(void)
-{
-#define ADD_I(name, nargs) add_intrinsic_properties(name, nargs, (void(*)(void))&jl_##name);
-#define ADD_HIDDEN ADD_I
-#define ALIAS(alias, base) add_intrinsic_properties(alias, intrinsic_nargs[base], runtime_fp[base]);
-    INTRINSICS
-#undef ADD_I
-#undef ADD_HIDDEN
-#undef ALIAS
-
-}
-
-#ifdef __cplusplus
-extern "C"
-#endif
-void jl_init_intrinsic_functions(void)
-{
-    jl_module_t *inm = jl_new_module(jl_symbol("Intrinsics"));
-    inm->parent = jl_core_module;
-    jl_set_const(jl_core_module, jl_symbol("Intrinsics"), (jl_value_t*)inm);
-    jl_mk_builtin_func(jl_intrinsic_type, "IntrinsicFunction", jl_f_intrinsic_call);
-
-#define ADD_I(name, nargs) add_intrinsic(inm, #name, name);
-#define ADD_HIDDEN(name, nargs)
-#define ALIAS ADD_I
-    INTRINSICS
-#undef ADD_I
-#undef ADD_HIDDEN
-#undef ALIAS
-}
